+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/BootstrapServer.lua
@@ -6,10 +6,22 @@
 --       SETTINGS
 --  ##################
 
+--  Default Settings
+--  ================
 S7_ConfigSettings = {
-    ["ConfigFiles"] = {"S7_Config.json", "S7_SecConfig.json"},
-    ["SyncStatPersistence"] = true
+    ["ConfigFiles"] = {"S7_Config.json"}, --  A list of all the files the configurator will pull from
+    ["SyncStatPersistence"] = true, --  Changes made with Ext.SyncStat() will be stored persistently if true
+    ["ExportStatIDtoTSV"] = {
+        ["FileName"] = "S7_Config_AllTheStats.tsv", --  FileName for ExportedStats
+        ["RestrictStatTypeTo"] = {} --  Limit the search to only these statTypes.
+    },
+    ["BypassSafetyCheck"] = false --  Bypasses S7_SafeToModify()
 }
+
+--  Import Custom Settings
+--  ======================
+
+--  <TODO>
 
 --  ##################
 --       MOD-MENU
@@ -17,22 +29,25 @@ S7_ConfigSettings = {
 
 local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris
     --  =====   STATS-CONFIGURATOR  =====
-    if Signal == "S7_StatsConfigurator" then --  Load JSON and Configure Stats
-        local files = S7_ConfigSettings.ConfigFiles
-        for i, fileName in ipairs(files) do
+    if Signal == "S7_StatsConfigurator" then --  Player requests stats-configuration
+        local files = S7_ConfigSettings.ConfigFiles --  list of all config files
+        for i, fileName in ipairs(files) do --  Iterate over each file
             Ext.Print("[S7:Config - BootstrapServer.lua] --- Loading " .. fileName)
             local JSONstring = Ext.LoadFile(fileName) --  Loads Configuration File
-            S7_StatsConfigurator(JSONstring) --  Call StatsConfigurator
+            S7_StatsConfigurator(JSONstring) --  Calls StatsConfigurator
         end
     end
-    --  =====   STATS-SYNCHRONIZE   =====    IS THIS EVEN NEEDED ?
+    --  =====   STATS-SYNCHRONIZE   =====    IS THIS EVEN NEEDED ? Completely useless as of now.
     if Signal == "S7_StatsSynchronize" then --  Synchronize stats between all clients
         Ext.Print("[S7:Config - BootstrapServer.lua] --- Synchronizing at Player's request.")
         S7_StatsSynchronize() --  Call StatsSynchronize
     end
     --  =====   EXPORT STATS TO TSV   =====
     if Signal == "S7_StatsExportTSV" then --  Export stat-types and stat-names to a tsv
-        Ext.Print("[S7:Config - BootstrapServer.lua] --- Exporting all StatIDs to S7_Config_AllTheStats.tsv")
+        Ext.Print(
+            "[S7:Config - BootstrapServer.lua] --- Exporting StatIDs to " ..
+                S7_ConfigSettings.ExportStatIDtoTSV.FileName
+        )
         S7_StatsExportTSV() --  Logs statIDs in an external TSV file for reference
     end
 end
@@ -48,10 +63,10 @@ Ext.NewCall(S7_Config_ModMenuRelay, "S7_Config_ModMenuRelay", "(STRING)_Signal")
 toSync = {} --  will hold a list of stats that were modified. for Ext.SyncStat()
 
 function S7_StatsConfigurator(JSONstring) --  Recieves stringified JSON from Ext.LoadFile(), Osiris or a mod.
-    local JSONborne = Ext.JsonParse(JSONstring) --  Parse JSONstring.
+    local JSONborne = Ext.JsonParse(JSONstring) --  Parsed JSONstring.
 
     if JSONborne ~= nil then --  JSONborne is not empty.
-        Ext.Print("[S7:Config - BootstrapServer.lua] --- JSON loaded. Applying Configuration Profile.\n\n")
+        Ext.Print("[S7:Config - BootstrapServer.lua] --- JSON loaded. Applying Configuration Profile.\n")
 
         Ext.Print("=============================================================")
         for name, content in pairs(JSONborne) do --  Iterate over JSONborne
@@ -61,10 +76,11 @@ function S7_StatsConfigurator(JSONstring) --  Recieves stringified JSON from Ext
             local stat = Ext.GetStat(name) --  Gets original stat-entry
             for key, value in pairs(content) do
                 if S7_SafeToModify(key) then --  Checks if key is safe to modify
-                    Ext.Print(key .. ": " .. value .. "\t(" .. stat[key] .. ")") --  e.g. - ActionPoints: 5(2)   |   StatName: NewValue(OriginalValue)
+                    Ext.Print(key .. ": " .. value .. " (" .. stat[key] .. ")") --  e.g. - ActionPoints: 5(2)   |   StatName: NewValue(OriginalValue)
                     stat[key] = value --  Sets new value for Name[Attribute]
                 end
             end
+
             Ext.Print("_____________________________________________________________")
             Ext.Print("\n")
             table.insert(toSync, name) --  Records stat-ids of the modified stats. To call Ext.SyncStat() on them later.
@@ -79,11 +95,13 @@ end
 
 function S7_StatsSynchronize()
     if type(next(toSync)) ~= "nil" then --  Stats were modified.
-        Ext.Print("[S7:Config - BootstrapServer.lua] --- Synchronizing Stats")
+        Ext.Print(
+            "[S7:Config - BootstrapServer.lua] --- Synchronizing [Persistence: " ..
+                S7_ConfigSettings.SyncStatPersistence .. "]"
+        )
         Ext.Print("=============================================================")
 
         for i, name in ipairs(toSync) do
-            Ext.Print(S7_ConfigSettings.SyncStatPersistence)
             Ext.SyncStat(name, S7_ConfigSettings.SyncStatPersistence) --  Sync
             Ext.Print("Synchronized Stat: " .. name)
             toSync[i] = nil --  Clears out toSync entry.
@@ -99,14 +117,25 @@ end
 --  ################################
 
 function S7_StatsExportTSV() --  Fetches literally every stat and exports to TSV
-    Ext.SaveFile("S7_Config_AllTheStats.tsv", "") --  Creates an empty TSV or Overwrites the existing one.
+    Ext.SaveFile(S7_ConfigSettings.ExportStatIDtoTSV.FileName, "") --  Creates an empty TSV or Overwrites the existing one.
     local SaveAllStatsToFile = "S.No\tType\tStatID\n" --  Header Column
-    local allStat = Ext.GetStatEntries() --  Get All Stat Entries
-    for key, value in ipairs(allStat) do
-        local type = NRD_StatGetType(value)
-        SaveAllStatsToFile = SaveAllStatsToFile .. key .. "\t" .. type .. "\t" .. value .. "\n"
+
+    local allStat = {} --  Initialize temporary table
+    if type(next(S7_ConfigSettings.ExportStatIDtoTSV.RestrictStatTypeTo)) == "nil" then --  No restrictions in settings
+        allStat = Ext.GetStatEntries() --  Get All Stat Entries
+    else
+        for i, statType in ipairs(S7_ConfigSettings.ExportStatIDtoTSV.RestrictStatTypeTo) do --  Only selected statTypes are loaded
+            table.insert(allStat, Ext.GetStatEntries(statType)) --  appends selected stat-type entries to allStat
+        end
     end
-    Ext.SaveFile("S7_Config_AllTheStats.tsv", SaveAllStatsToFile) --  Save TSV file.
+
+    for i, limitedAllStat in ipairs(allStat) do
+        for key, value in ipairs(limitedAllStat) do --  Iterate over allStat
+            local type = NRD_StatGetType(value) --  Didn't I just filter stats by statType? - this is what happens when you return to old code with new ideas
+            SaveAllStatsToFile = SaveAllStatsToFile .. key .. "\t" .. type .. "\t" .. value .. "\n" --  Tab Separated Values
+        end
+    end
+    Ext.SaveFile(S7_ConfigSettings.ExportStatIDtoTSV.FileName, SaveAllStatsToFile) --  Save TSV file.
 end
 
 --  ###########################
@@ -115,6 +144,7 @@ end
 
 local function S7_InspectStats(StatID, StatType) --  Recieves StatID and StatType from Osiris
     local compareStat = Ext.GetStatEntries(StatType) --  Retrieves all stat entries of corresponding stat-type for comparison.
+    Ext.Print("[S7:Config - BootstrapServer.lua] --- Inspecting Target")
     for name, content in pairs(compareStat) do --  Iterate over compareStat
         if content == StatID then
             Ext.Print("[S7:Config - BootstrapServer.lua] --- (" .. StatType .. "): " .. StatID)
@@ -122,9 +152,9 @@ local function S7_InspectStats(StatID, StatType) --  Recieves StatID and StatTyp
     end
 end
 
---  ==================================================================================
+--  =================================================================================
 Ext.NewCall(S7_InspectStats, "S7_InspectStats", "(STRING)_StatID, (STRING)_StatType")
---  ==================================================================================
+--  =================================================================================
 
 --  ############################
 --      FUNCTION DEFINITIONS
@@ -144,11 +174,15 @@ function S7_SafeToModify(key) --  Checks if key is safe to modify.
         "LoseBoost"
     }
 
-    for i, avoid in ipairs(dontFwith) do
-        if key == avoid then
-            return false
-        else
-            return true
+    if S7_ConfigSettings.BypassSafetyCheck == true then --  Manual-Override setting is true
+        return true --  S7_SafeToModify() returns true for everything
+    else -- Default Setting
+        for i, avoid in ipairs(dontFwith) do --  Iterate over keys to avoid
+            if key == avoid then --  If key matches
+                return false --  Stop it right there
+            else
+                return true --  else continue
+            end
         end
     end
 end